server/requirements.txt
flask
flask-cors
python-dotenv
torch
transformers
diffusers
accelerate
sentence-transformers
faiss-cpu
opencv-python
pyjwt
bcrypt
pymongo